% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_gen}
\alias{data_gen}
\title{Generate simulated dataset}
\usage{
data_gen(n, noncompliance = "two-sided")
}
\arguments{
\item{n}{sample size.}

\item{noncompliance}{"one-sided" or "two-sided" noncompliance type}
}
\value{
A simulated data with sample size n.
}
\description{
Generate simulated dataset
}
\examples{
head(data_gen(n=100))
}
