% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GMF_nonpar}
\alias{GMF_nonpar}
\title{Nonparametric efficient estimator for calculating the generalized mediation functional theta_(d1d0)^(zz')}
\usage{
GMF_nonpar(
  Xpi,
  Xe,
  Xm,
  Xo,
  Z,
  D,
  M,
  Y,
  group = "compliers",
  type = "10",
  noncompliance = "two-sided",
  learners = c("SL.glm", "SL.ranger"),
  V = 5,
  output.IF = FALSE
)
}
\arguments{
\item{Xpi}{design matrix for the treatment probability model}

\item{Xe}{design matrix for the principal score model}

\item{Xm}{design matrix for the mediator model}

\item{Xo}{design matrix for the outcome model}

\item{Z}{vector for the treatment}

\item{D}{vector for the post-treatment event}

\item{M}{vector for the mediator}

\item{Y}{vector for the outcome}

\item{group}{principal stratum ("compliers", "always takers", or "never takers")}

\item{type}{superscript for the mediation functional (i.e., zz'). The default value is "10".}

\item{noncompliance}{noncompliance type (one-sided or two-sided corresponding to strong and standard monotonicity assumption)}

\item{learners}{names of super learners (default GLM and random forest)}

\item{V}{number of folds of the cross-fitting procedure (default 5)}

\item{output.IF}{whether output the influence function (default false)}
}
\value{
estimates for the generalized mediation functional based on the nonparametric efficient estimator
}
\description{
Nonparametric efficient estimator for calculating the generalized mediation functional theta_(d1d0)^(zz')
}
