% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GMF_par}
\alias{GMF_par}
\title{Singly and multiply robust estimators for calculating the generalized mediation functional theta_(d1d0)^(zz')}
\usage{
GMF_par(
  Xpi,
  Xe,
  Xm,
  Xo,
  Z,
  D,
  M,
  Y,
  group = "compliers",
  type = "10",
  noncompliance = "two-sided"
)
}
\arguments{
\item{Xpi}{design matrix for the treatment probability model}

\item{Xe}{design matrix for the principal score model}

\item{Xm}{design matrix for the mediator model}

\item{Xo}{design matrix for the outcome model}

\item{Z}{vector for the treatment}

\item{D}{vector for the post-treatment event}

\item{M}{vector for the mediator}

\item{Y}{vector for the outcome}

\item{group}{principal stratum ("compliers", "always takers", or "never takers")}

\item{type}{superscript for the mediation functional (i.e., zz'). The default value is "10".}

\item{noncompliance}{noncompliance type (one-sided or two-sided corresponding to strong and standard monotonicity assumption)}
}
\value{
estimates for the generalized mediation functional based on the singly robust estimator as
well as the multiply robust estimator
}
\description{
Singly and multiply robust estimators for calculating the generalized mediation functional theta_(d1d0)^(zz')
}
