% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_np}
\alias{mediate_np}
\title{Nonparametric efficient estimation of the principal natural (in)direct effects}
\usage{
mediate_np(
  data,
  Xname,
  Yname,
  Zname,
  Dname,
  Mname,
  monotonicity = "standard",
  learners = c("SL.glm", "SL.ranger"),
  V = 5
)
}
\arguments{
\item{data}{dataset}

\item{Xname}{names of the baseline covariates}

\item{Yname}{name of the outcome}

\item{Zname}{name of the treatment}

\item{Dname}{name of the post-treatment event}

\item{Mname}{name of the mediator}

\item{monotonicity}{"strong" or "standard" monotonicity assumption (default "standard")}

\item{learners}{names of super learners (default GLM and random forest)}

\item{V}{number of folds for the cross-fitting procedure (default 5)}
}
\value{
Point, SE, and 95% confidence interval estimates of the principal natural (in)direct effects
}
\description{
Nonparametric efficient estimation of the principal natural (in)direct effects
}
\examples{
# set.seed(2024)
# data = data_gen(n=500)
# mediate_np(data,
#            Xname=c("X1","X2","X3","X4"),
#            Yname="Y",
#            Zname="Z",
#            Dname="D",
#            Mname="M",
#            monotonicity="standard",
#            learners=c("SL.glm", "SL.ranger"),V=3)

}
