% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_par}
\alias{mediate_par}
\title{Singly and multiply robust estimation of the principal natural (in)direct effects}
\usage{
mediate_par(
  data,
  estimator = "a",
  Xname,
  Yname,
  Zname,
  Dname,
  Mname,
  monotonicity = "standard",
  B = 500
)
}
\arguments{
\item{data}{dataset}

\item{estimator}{For singly robust estimation, set it to "a", "b", "c", "d" for estimator a, b, c, d, respectively. For multiply robust estimation, set it to "mr"}

\item{Xname}{names of the baseline covariates}

\item{Yname}{name of the outcome}

\item{Zname}{name of the treatment}

\item{Dname}{name of the post-treatment event}

\item{Mname}{name of the mediator}

\item{monotonicity}{"strong" or "standard" monotonicity assumption (default "standard")}

\item{B}{number of iterations for the nonparametric bootstrap procedure}
}
\value{
Point, SE, and 95% confidence interval estimates of the principal natural (in)direct effects
}
\description{
Singly and multiply robust estimation of the principal natural (in)direct effects
}
\examples{
# set.seed(2024)
# data = data_gen(n=500)
# mediate_par(data,
#             estimator="mr",
#             Xname=c("X1","X2","X3","X4"),
#             Yname="Y",
#             Zname="Z",
#             Dname="D",
#             Mname="M",
#             monotonicity="standard",
#             B=200)
}
