% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mediate_par_point}
\alias{mediate_par_point}
\title{Point estimates of the principal natural (in)direct effects based on the singly and multiply robust estimators}
\usage{
mediate_par_point(
  data,
  Xname,
  Yname,
  Zname,
  Dname,
  Mname,
  monotonicity = "standard"
)
}
\arguments{
\item{data}{dataset}

\item{Xname}{names of the baseline covariates}

\item{Yname}{name of the outcome}

\item{Zname}{name of the treatment}

\item{Dname}{name of the post-treatment event}

\item{Mname}{name of the mediator}

\item{monotonicity}{"strong" or "standard" monotonicity assumption (default "standard")}
}
\value{
Point estimate of the principal natural (in)direct effects
}
\description{
Point estimates of the principal natural (in)direct effects based on the singly and multiply robust estimators
}
